US2025072402 Play And Recd Function, analysis_config_window.py, acquisition_config_window.py
